+++ app/models/__init__.py
@@ -4,3 +4,4 @@ from .upload_file import UploadFile
 from .equipe import Equipe
 from .circulo import Circulo
 from .encontreiro import Encontreiro
+from .encontrista import Encontrista